chore: Rework directory structure
- Have both pages and app implementations in parallel
- Expose named exports (default links to pages for compat)
- Separate e2e pages for individual testing
- Separate tsd test files for individual API testing (not ideal)
