+++ src/app/app/useQueryState/page.tsx
@@ -1,7 +1,6 @@
 'use client'
 
-import React from 'react'
-import { queryTypes, useQueryState } from './../../../../'
+import { queryTypes, useQueryState } from '../../../../dist/app'
 
 const IntegrationPage = () => {
   const [string, setString] = useQueryState('string')
