+++ src/app/app/useQueryStates/dynamic/[route]/page.tsx
@@ -1,0 +1,4 @@
+'use client'
+
+import IntegrationPage from '../../page'
+export default IntegrationPage
