+++ src/app/layout.tsx
@@ -1,8 +1,10 @@
-'use client'
-
 import React from 'react'
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function RootLayout({
+  children
+}: {
+  children: React.ReactNode
+}) {
   return (
     <html>
       <body>{children}</body>
