--- src/index.ts
@@ -1,3 +1,0 @@
-export * from './defs'
-export * from './useQueryState'
-export * from './useQueryStates'
