+++ src/lib/app/defs.ts
@@ -1,0 +1,1 @@
+export * from '../commonDefs'
