+++ src/lib/app/useQueryState.ts
@@ -1,6 +1,6 @@
-import { useRouter } from 'next/router'
+import { usePathname, useRouter, useSearchParams } from 'next/navigation'
 import React from 'react'
-import { HistoryOptions, Serializers, TransitionOptions } from './defs'
+import type { HistoryOptions, Serializers } from '../commonDefs'
 
 export interface UseQueryStateOptions<T> extends Serializers<T> {
   /**
@@ -18,10 +18,9 @@ export type UseQueryStateReturn<Parsed, Default> = [
       | null
       | Parsed
       | ((
-      old: Default extends Parsed ? Parsed : Parsed | null
-    ) => Parsed | null),
-    transitionOptions?: TransitionOptions
-  ) => Promise<boolean>
+          old: Default extends Parsed ? Parsed : Parsed | null
+        ) => Parsed | null)
+  ) => void
 ]
 
 // Overload type signatures ----------------------------------------------------
@@ -205,6 +204,8 @@ export function useQueryState<T = string>(
   }
 ) {
   const router = useRouter()
+  const pathname = usePathname()
+  const searchParams = useSearchParams()
 
   // Memoizing the update function has the advantage of making it
   // immutable as long as `history` stays the same.
@@ -214,26 +215,29 @@ export function useQueryState<T = string>(
     [history]
   )
 
-  const getValue = React.useCallback((): T | null => {
+  function getValue(mode: 'current' | 'previous' = 'current'): T | null {
     if (typeof window === 'undefined') {
       // Not available in an SSR context
       return null
     }
-    const query = new URLSearchParams(window.location.search)
-    const value = query.get(key)
-    return value !== null ? parse(value) : null
-  }, [])
+    const query =
+      mode === 'current'
+        ? searchParams
+        : new URLSearchParams(window.location.search)
+    const value = query?.get(key)
+    return value != null ? parse(value) : null
+  }
 
-  // Update the state value only when the relevant key changes.
-  // Because we're not calling getValue in the function argument
-  // of React.useMemo, but instead using it as the function to call,
-  // there is no need to pass it in the dependency array.
-  const value = React.useMemo(getValue, [router.query[key]])
+  const getOldValue = React.useCallback(
+    (): T | null => getValue('previous'),
+    []
+  )
+
+  const value = getValue()
 
   const update = React.useCallback(
     (
-      stateUpdater: React.SetStateAction<T | null>, // todo: This is wrong now
-      transitionOptions?: TransitionOptions
+      stateUpdater: React.SetStateAction<T | null> // todo: This is wrong now
     ) => {
       const isUpdaterFunction = (
         input: any
@@ -242,38 +246,25 @@ export function useQueryState<T = string>(
       }
 
       // Resolve the new value based on old value & updater
-      const oldValue = getValue() ?? defaultValue ?? null
+      const oldValue = getOldValue() ?? defaultValue ?? null
       const newValue = isUpdaterFunction(stateUpdater)
         ? stateUpdater(oldValue)
         : stateUpdater
-      // We can't rely on router.query here to avoid causing
-      // unnecessary renders when other query parameters change.
       // URLSearchParams is already polyfilled by Next.js
-      const query = new URLSearchParams(window.location.search)
+      const query = new URLSearchParams(searchParams?.toString())
       if (newValue === null) {
         // Don't leave value-less keys hanging
         query.delete(key)
-      } else {
+      } else if (newValue !== undefined) {
         query.set(key, serialize(newValue))
       }
       const search = query.toString()
       const hash = window.location.hash
-      return updateUrl?.call(
-        router,
-        {
-          pathname: router.pathname,
-          hash,
-          search
-        },
-        {
-          pathname: window.location.pathname,
-          hash,
-          search
-        },
-        transitionOptions
-      )
+
+      updateUrl?.call(router, `${pathname}${search ? '?' + search : ''}${hash}`)
     },
-    [key, updateUrl]
+    [key, searchParams, updateUrl]
   )
+
   return [value ?? defaultValue ?? null, update]
 }
