+++ src/lib/app/useQueryStates.ts
@@ -1,11 +1,6 @@
-import { useRouter } from 'next/router'
+import { usePathname, useRouter, useSearchParams } from 'next/navigation'
 import React from 'react'
-import type {
-  HistoryOptions,
-  Nullable,
-  Serializers,
-  TransitionOptions
-} from './defs'
+import type { HistoryOptions, Nullable, Serializers } from '../commonDefs'
 
 type KeyMapValue<Type> = Serializers<Type> & {
   defaultValue?: Type
@@ -24,8 +19,8 @@ export interface UseQueryStatesOptions {
 
 export type Values<T extends UseQueryStatesKeysMap> = {
   [K in keyof T]: T[K]['defaultValue'] extends NonNullable<
-      ReturnType<T[K]['parse']>
-    >
+    ReturnType<T[K]['parse']>
+  >
     ? NonNullable<ReturnType<T[K]['parse']>>
     : ReturnType<T[K]['parse']> | null
 }
@@ -35,9 +30,8 @@ type UpdaterFn<T extends UseQueryStatesKeysMap> = (
 ) => Partial<Nullable<Values<T>>>
 
 export type SetValues<T extends UseQueryStatesKeysMap> = (
-  values: Partial<Nullable<Values<T>>> | UpdaterFn<T>,
-  transitionOptions?: TransitionOptions
-) => Promise<boolean>
+  values: Partial<Nullable<Values<T>>> | UpdaterFn<T>
+) => void
 
 export type UseQueryStatesReturn<T extends UseQueryStatesKeysMap> = [
   Values<T>,
@@ -50,12 +44,15 @@ export type UseQueryStatesReturn<T extends UseQueryStatesKeysMap> = [
  * @param keys - An object describing the keys to synchronise and how to
  *               serialise and parse them.
  *               Use `queryTypes.(string|integer|float)` for quick shorthands.
+ * @param options - Optional history mode.
  */
 export function useQueryStates<KeyMap extends UseQueryStatesKeysMap>(
   keys: KeyMap,
   { history = 'replace' }: Partial<UseQueryStatesOptions> = {}
 ): UseQueryStatesReturn<KeyMap> {
   const router = useRouter()
+  const pathname = usePathname()
+  const searchParams = useSearchParams()
 
   type V = Values<KeyMap>
 
@@ -67,7 +64,7 @@ export function useQueryStates<KeyMap extends UseQueryStatesKeysMap>(
     [history]
   )
 
-  const getValues = React.useCallback((): V => {
+  const getValues = (mode: 'current' | 'previous' = 'current'): V => {
     if (typeof window === 'undefined') {
       // Not available in an SSR context, return all null (or default if available)
       return Object.keys(keys).reduce((obj, key) => {
@@ -78,12 +75,15 @@ export function useQueryStates<KeyMap extends UseQueryStatesKeysMap>(
         }
       }, {} as V)
     }
-    const query = new URLSearchParams(window.location.search)
+    const query =
+      mode === 'current'
+        ? searchParams
+        : new URLSearchParams(window.location.search)
     return Object.keys(keys).reduce((values, key) => {
       const { parse, defaultValue } = keys[key as keyof KeyMap]
-      const value = query.get(key)
+      const value = query?.get(key)
       const parsed =
-        value !== null
+        value != null
           ? parse(value) ?? defaultValue ?? null
           : defaultValue ?? null
       return {
@@ -91,32 +91,25 @@ export function useQueryStates<KeyMap extends UseQueryStatesKeysMap>(
         [key]: parsed
       }
     }, {} as V)
-  }, [keys])
+  }
 
-  // Update the state values only when the relevant keys change.
-  // Because we're not calling getValues in the function argument
-  // of React.useMemo, but instead using it as the function to call,
-  // there is no need to pass it in the dependency array.
-  const values = React.useMemo(
-    getValues,
-    Object.keys(keys).map(key => router.query[key])
-  )
+  const getOldValues = React.useCallback((): V => getValues('previous'), [])
+
+  const values = getValues()
 
   const update = React.useCallback<SetValues<KeyMap>>(
-    (stateUpdater, transitionOptions) => {
+    stateUpdater => {
       const isUpdaterFunction = (input: any): input is UpdaterFn<KeyMap> => {
         return typeof input === 'function'
       }
 
       // Resolve the new values based on old values & updater
-      const oldValues = getValues()
+      const oldValues = getOldValues()
       const newValues = isUpdaterFunction(stateUpdater)
         ? stateUpdater(oldValues)
         : stateUpdater
-      // We can't rely on router.query here to avoid causing
-      // unnecessary renders when other query parameters change.
       // URLSearchParams is already polyfilled by Next.js
-      const query = new URLSearchParams(window.location.search)
+      const query = new URLSearchParams(searchParams?.toString())
 
       Object.keys(newValues).forEach(key => {
         const newValue = newValues[key]
@@ -129,22 +122,9 @@ export function useQueryStates<KeyMap extends UseQueryStatesKeysMap>(
       })
       const search = query.toString()
       const hash = window.location.hash
-      return updateUrl?.call(
-        router,
-        {
-          pathname: router.pathname,
-          hash,
-          search
-        },
-        {
-          pathname: window.location.pathname,
-          hash,
-          search
-        },
-        transitionOptions
-      )
+      updateUrl?.call(router, `${pathname}${search ? '?' + search : ''}${hash}`)
     },
-    [keys, updateUrl]
+    [keys, searchParams, updateUrl]
   )
   return [values, update]
 }
