+++ src/lib/commonDefs.ts
