+++ src/lib/index.app.ts
@@ -1,0 +1,6 @@
+'use client'
+
+export * from './app/useQueryState'
+export * from './app/useQueryStates'
+export * from './commonDefs'
+export const __useQueryStateRouter = 'app'
