+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+'use client'
+
+export * from './index.pages'
