+++ src/lib/pages/defs.ts
@@ -1,0 +1,5 @@
+import type { Router } from 'next/router'
+
+// Next.js does not export the TransitionsOption interface,
+// but we can get it from where it's used:
+export type TransitionOptions = Parameters<Router['push']>[2]
