+++ src/pages/pages/useQueryStates/dynamic/[route]/index.tsx
@@ -1,0 +1,4 @@
+'use client'
+
+import IntegrationPage from '../../index'
+export default IntegrationPage
