+++ src/tests/pages/useQueryState.test-d.ts
@@ -1,5 +1,5 @@
 import { expectError, expectNotAssignable, expectType } from 'tsd'
-import { queryTypes, useQueryState } from '../../legacy/index'
+import { queryTypes, useQueryState } from '../../../dist/pages'
 
 // By default, queries have a `string` state, nullable (when no query parameter is present)
 {
